opt: improve the internal packages of math and bytes
