--- internal/toolkit/math.go
@@ -1,52 +1,0 @@
-// Copyright (c) 2017-2019 Sergey Kamardin <gobwas@gmail.com>
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file.
-
-package toolkit
-
-const (
-	bitsize       = 32 << (^uint(0) >> 63)
-	maxintHeadBit = 1 << (bitsize - 2)
-)
-
-// IsPowerOfTwo reports whether given integer is a power of two.
-func IsPowerOfTwo(n int) bool {
-	return n&(n-1) == 0
-}
-
-// CeilToPowerOfTwo returns the least power of two integer value greater than
-// or equal to n.
-func CeilToPowerOfTwo(n int) int {
-	if n&maxintHeadBit != 0 && n > maxintHeadBit {
-		panic("argument is too large")
-	}
-	if n <= 2 {
-		return 2
-	}
-	n--
-	n = fillBits(n)
-	n++
-	return n
-}
-
-// FloorToPowerOfTwo returns the greatest power of two integer value less than
-// or equal to n.
-func FloorToPowerOfTwo(n int) int {
-	if n <= 2 {
-		return 2
-	}
-	n = fillBits(n)
-	n >>= 1
-	n++
-	return n
-}
-
-func fillBits(n int) int {
-	n |= n >> 1
-	n |= n >> 2
-	n |= n >> 4
-	n |= n >> 8
-	n |= n >> 16
-	n |= n >> 32
-	return n
-}
